backend/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
requests==2.32.3
python-dotenv==1.0.1
SQLAlchemy==2.0.36
pydantic>=2.11.7,<3
numpy==2.1.3
pandas==2.2.3
scikit-learn==1.5.2
joblib==1.4.2
